+++ src/public/InvokeNovaTest.ps1
@@ -10,16 +10,27 @@ function Invoke-NovaTest {
     )
 
     dynamicparam {
-        return Get-NovaDynamicOverrideWarningParameterDictionary
+        $dictionary = Get-NovaDynamicOverrideWarningParameterDictionary
+        $attributeCollection = [System.Collections.ObjectModel.Collection[System.Attribute]]::new()
+        $attributeCollection.Add([System.Management.Automation.ParameterAttribute]::new())
+        $runtimeParameter = [System.Management.Automation.RuntimeDefinedParameter]::new('PesterConfigurationOverride', [hashtable], $attributeCollection)
+        $dictionary.Add('PesterConfigurationOverride', $runtimeParameter)
+        return $dictionary
     }
 
     end {
+        $pesterConfigurationOverride = $null
+        if ($PSBoundParameters.ContainsKey('PesterConfigurationOverride')) {
+            $pesterConfigurationOverride = [hashtable]$PSBoundParameters.PesterConfigurationOverride
+        }
+
         $workflowContext = Get-NovaTestWorkflowContext -TestOption @{
             TestMode = 'Unit'
             TagFilter = $TagFilter
             ExcludeTagFilter = $ExcludeTagFilter
             OutputVerbosity = $OutputVerbosity
             OutputRenderMode = $OutputRenderMode
+            PesterConfigurationOverride = $pesterConfigurationOverride
         } -BoundParameters $PSBoundParameters
 
         $shouldRun = $PSCmdlet.ShouldProcess($workflowContext.Target, $workflowContext.Operation)